Warning fixes
git-svn-id: https://svn.remlab.net/svn/miredo/trunk@208 6046fb1f-dae2-0310-8472-869a43925389

+++ src/miredo.cpp
@@ -471,7 +471,8 @@ miredo_run (uint16_t bind_port, const char *bind_ip, const char *server_name,
 	{
 		syslog (LOG_ALERT,
 			_("Teredo service port failure: "
-			"cannot open UDP port %u\n"), ntohs (bind_port));
+			"cannot open UDP port %u\n"),
+			(unsigned int)ntohs (bind_port));
 		syslog (LOG_NOTICE, _("Make sure another instance "
 			"of the program is not already running."));
 		goto abort;
